third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/script-tag-expression-follows-expected.txt

CONSOLE ERROR: line 5: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?script-expression-follows=1&q=%3Cscript%3Ealert(%27XSS%27)' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
